# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API BTM",
        "description": "API du site BTM",
        "contact": {
            "email": "contact@balancetonmatch.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api",
            "description": "Demo API Server"
        }
    ],
    "paths": {
...
        "/clubs/{clubId}/uploadImage": {
            "post": {
                "tags": [
                    "club"
                ],
                "summary": "uploads an image",
                "operationId": "uploadFile",
                "parameters": [
                    {
                        "name": "clubId",
                        "in": "path",
                        "description": "ID of club to update",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "description": "Upload images request body",
                    "content": {
                        "application/octet-stream": {
                            "schema": {
                                "type": "string",
                                "format": "binary"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "http://127.0.0.1:8000/components/schemas/ApiResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "clubstore_auth": [
                            "write:pets",
                            "read:pets"
                        ]
                    }
                ]
            }
        },
        "/comments": {
            "put": {
                "tags": [
                    "comment"
                ],
                "summary": "Update an existing comment",
                "operationId": "updateComment",
                "requestBody": {
                    "$ref": "http://127.0.0.1:8000/components/requestBodies/Comment"
                },
                "responses": {
                    "400": {
                        "description": "Invalid ID supplied"
                    },
                    "404": {
                        "description": "Comment not found"
                    },
                    "405": {
                        "description": "Validation exception"
                    }
                },
                "security": [
                    {
                        "commentstore_auth": [
                            "write:comments",
                            "read:comments"
                        ]
                    }
                ]
            },
            "post": {
                "tags": [
                    "comment"
                ],
                "summary": "Add a new comment to the store",
                "operationId": "addComment",
                "requestBody": {
                    "$ref": "http://127.0.0.1:8000/components/requestBodies/Comment"
                },
                "responses": {
                    "405": {
                        "description": "Invalid input"
                    }
                },
                "security": [
                    {
                        "commentstore_auth": [
                            "write:comments",
                            "read:comments"
                        ]
                    }
                ]
            }
        },
        "/comments/{commentId}": {
            "get": {
                "tags": [
                    "comment"
                ],
                "summary": "Find comment by ID",
                "description": "Returns a single comment",
                "operationId": "getCommentById",
                "parameters": [
                    {
                        "name": "commentId",
                        "in": "path",
                        "description": "ID of comment to return",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "http://127.0.0.1:8000/components/schemas/Comment"
                                }
                            },
                            "application/xml": {
                                "schema": {
                                    "$ref": "http://127.0.0.1:8000/components/schemas/Comment"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid ID supplier"
                    },
                    "404": {
                        "description": "Comment not found"
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            }
        },
        "/comment/{commentId}": {
            "post": {
                "tags": [
                    "comment"
                ],
                "summary": "Updates a comment in the store with form data",
                "operationId": "updateCommentWithForm",
                "parameters": [
                    {
                        "name": "commentId",
                        "in": "path",
                        "description": "ID of comment that needs to be updated",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Input data format",
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "description": "Updated name of the comment",
                                        "type": "string"
                                    },
                                    "status": {
                                        "description": "Updated status of the comment",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "405": {
                        "description": "Invalid input"
                    }
                },
                "security": [
                    {
                        "commentstore_auth": [
                            "write:comments",
                            "read:comments"
                        ]
                    }
                ]
            },
            "delete": {
                "tags": [
                    "comment"
                ],
                "summary": "Deletes a comment",
                "operationId": "deleteComment",
                "parameters": [
                    {
                        "name": "api_key",
                        "in": "header",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "commentId",
                        "in": "path",
                        "description": "Comment id to delete",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "400": {
                        "description": "Invalid ID supplied"
                    },
                    "404": {
                        "description": "Comment not found"
                    }
                },
                "security": [
                    {
                        "commentstore_auth": [
                            "write:comments",
                            "read:comments"
                        ]
                    }
                ]
            }
        },
        "/comment/{commentId}/uploadImage": {
            "post": {
                "tags": [
                    "comment"
                ],
                "summary": "uploads an image",
                "operationId": "uploadFile",
                "parameters": [
                    {
                        "name": "commentId",
                        "in": "path",
                        "description": "ID of comment to update",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "description": "Upload images request body",
                    "content": {
                        "application/octet-stream": {
                            "schema": {
                                "type": "string",
                                "format": "binary"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "http://127.0.0.1:8000/components/schemas/ApiResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "commentstore_auth": [
                            "write:pets",
                            "read:pets"
                        ]
                    }
                ]
            }
        },
...
        "/user/{userId}/uploadImage": {
            "post": {
                "tags": [
                    "user"
                ],
                "summary": "uploads an image",
                "operationId": "uploadFile",
                "parameters": [
                    {
                        "name": "userId",
                        "in": "path",
                        "description": "ID of user to update",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "description": "Upload images request body",
                    "content": {
                        "application/octet-stream": {
                            "schema": {
                                "type": "string",
                                "format": "binary"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "http://127.0.0.1:8000/components/schemas/ApiResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "userstore_auth": [
                            "write:pets",
                            "read:pets"
                        ]
                    }
                ]
            }
        }
    },
    "security": [
        []
    ]
}